FIX: 폴더구조 변경 (rules, exception)
