+++ src/main/kotlin/kr/hhplus/be/server/domain/balance/models/PointHistory.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.domain.balance.models
 import kr.hhplus.be.server.global.common.BaseEntity
 import jakarta.persistence.*
 import kr.hhplus.be.server.domain.balance.exception.InvalidAmountException
-import kr.hhplus.be.server.domain.common.BalanceBusinessRules
+import kr.hhplus.be.server.domain.balance.rules.BalanceBusinessRules
 import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
