+++ src/main/kotlin/kr/hhplus/be/server/global/config/BusinessRulesConfig.kt
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.global.config
 
-import kr.hhplus.be.server.domain.common.ConcertBusinessRules
-import kr.hhplus.be.server.domain.common.ReservationBusinessRules
+import kr.hhplus.be.server.domain.concert.rules.ConcertBusinessRules
 import kr.hhplus.be.server.global.properties.ConcertProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
@@ -10,11 +9,6 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 @EnableConfigurationProperties(ConcertProperties::class)
 class BusinessRulesConfig {
-    
-    @Bean
-    fun reservationBusinessRules(concertProperties: ConcertProperties): ReservationBusinessRules {
-        return ReservationBusinessRules(concertProperties)
-    }
     
     @Bean
     fun concertBusinessRules(concertProperties: ConcertProperties): ConcertBusinessRules {
