+++ src/main/kotlin/kr/hhplus/be/server/global/exception/GlobalExceptionHandler.kt
@@ -1,12 +1,10 @@
 package kr.hhplus.be.server.global.exception
 
-import kr.hhplus.be.server.domain.common.DomainExceptionFactory
 import kr.hhplus.be.server.global.lock.ConcurrentAccessException
 import kr.hhplus.be.server.global.response.CommonApiResponse
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.validation.FieldError
 import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
